chore: separate createEngine from createHonoHandler
