+++ packages/waku/src/unstable_engine.ts
@@ -1,0 +1,4 @@
+// These exports are for internal use only and subject to change without notice.
+// Separated from unstable_hono and unstable_internals to not require node:fs.
+
+export { createEngine } from './vite-rsc/lib/engine.js';
