+++ packages/waku/src/vite-rsc/deploy/deno/entry.ts
@@ -4,7 +4,7 @@
 import { Hono } from 'jsr:@hono/hono';
 // @ts-expect-error deno
 import { serveStatic } from 'jsr:@hono/hono/deno';
-import { createHonoHandler } from '../../lib/engine.js';
+import { createHonoHandler } from '../../lib/hono.js';
 import { honoEnhancer } from 'virtual:vite-rsc-waku/hono-enhancer';
 import { config } from 'virtual:vite-rsc-waku/config';
 import path from 'node:path';
